UIの整理

+++ src/main/java/com/example/expense/api/Expense.java
@@ -9,7 +9,7 @@ public class Expense {
     private Integer id;
     private LocalDate date; // 日付
     private String title; // タイトル
-    private Double amount; // 金額
+    private Integer amount; // 金額
     private String payer; // 支払い者
     private Integer split; // 分割支払いかどうか
     private String category; // 費用カテゴリ（例：食費、交通費、など）
